@@ -6,7 +6,8 @@
   "scripts": {
     "start": "webpack-dev-server --mode development",
     "build": "webpack --mode production",
-    "pic": "webpack-dev-server --mode development --config packages/picture/webpack.config.js"
+    "pic": "webpack serve --mode development --config packages/picture/webpack.config.js",
+    "ts": "webpack serve --mode development --config packages/typescript/webpack.config.js"
   },
   "keywords": [],
   "author": "",
@@ -19,11 +20,18 @@
     "@babel/core": "^7.18.13",
     "@babel/preset-env": "^7.4.4",
     "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.18.6",
     "babel-loader": "^8.0.5",
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.1",
-    "webpack-dev-server": "^3.3.1",
+    "image-webpack-loader": "^8.1.0",
+    "style-loader": "^3.3.1",
+    "typescript": "^4.8.2",
+    "url-loader": "^4.1.1",
+    "webpack": "5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.10.0",
     "webpack-merge": "^5.8.0"
   }
 }
